+++ src/payloads/create_order_payload.rs
@@ -1,11 +1,13 @@
+use bigdecimal::BigDecimal;
 use serde::Deserialize;
 use crate::enums::{order_type::OrderType, types::OrderSide};
 
+#[allow(dead_code)]
 #[derive(Debug, Deserialize)]
 pub struct CreateOrderPayload {
     pub side: OrderSide,
     pub order_type: OrderType,
     pub symbol: String,
-    pub price: Option<u64>, 
-    pub quantity: u64,
+    pub price: Option<BigDecimal>, 
+    pub quantity: BigDecimal,
 }
