+++ src/payloads/order_payload.rs
@@ -1,3 +1,4 @@
+use bigdecimal::BigDecimal;
 use serde::Deserialize;
 
 use crate::enums::{order_type::OrderType, types::OrderSide};
@@ -7,6 +8,6 @@ pub struct CreateOrderPayload {
     pub side: OrderSide,
     pub order_type: OrderType,
     pub symbol: String,
-    pub price: Option<u64>,
-    pub quantity: u64,
+    pub price: Option<BigDecimal>,
+    pub quantity: BigDecimal,
 }
